base: portmgr/jobs/mirror_macports.sh: fix variant extraction code and loops
I'm not sure if this script is actively being used, but it certainly didn't
correctly extract variants (it ran, e.g. port mirror collectd +[+]ntpd).
Additionally, a loop over a quoted string in bash does not tokenize the
contents of the string and loop over the tokens, but loops over the whole thing
at once, so let's remove the quotes where we don't really want them.

git-svn-id: https://svn.macports.org/repository/macports/trunk/base/portmgr/jobs@135069 d073be05-634f-4543-b044-5fe20cf6d1d6