buildbot: Use mpbb list-subports, ignore stderr
Switch to ./mpbb list-subports over the older subports.tcl script. This has the
advantage of avoiding a loop, because that loop is now handled in mpbb.

Additionally, ignore output written to stderr and do not interpret it as port
name. Buildbot 0.8.x by default includes stderr while parsing the result and
the Portfiles we run in this step can print arbitrary strings to stderr.
Buildbot 0.9.x has a fix for this, but fortunately we can work around the
problem with the 0.8.x feature set as well.

git-svn-id: https://svn.macports.org/repository/macports/contrib/buildbot-test@151030 d073be05-634f-4543-b044-5fe20cf6d1d6