Add a basic mirroring step to buildbot cfg
Runs a trigger step in each ports-watcher before the one that triggers
the actual port builds. The exact command will need to be refined
depending on the exact conditions on the buildslave. It might need to
update a ports tree to work from, or run a port command installed
somewhere non-standard, etc.