containers/paste: Send status output to stdout
We're running deploy.sh under cronie, which now always sends an email
because there was output on stderr since docker has decided to always
write output there.

Fix that by redirecting the output to stdout, so that mails are only
sent when other commands print on stderr, or a command fails.