mprsyncup-alt: get latest ports commit from redis
This ensures that this job and the one running on GitHub Actions are
always on the same page about which ports tarballs can safely be
deleted. It's possible for Actions to upload a tarball but then fail
before it updates mirror.last_commit in the db, in which case it will
want to download a slightly older tarball in the next run.